Move up common remote datastore classes
