+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/NotFoundException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.gnd.persistence.remote.firestore;
+package com.google.android.gnd.persistence.remote;
 
 /** Exception thrown when a remote entity is not found. */
 public class NotFoundException extends DataStoreException {
